added cloudinary and working as background on card

+++ db/seeds.rb
@@ -1,13 +1,15 @@
 require "json"
 require 'open-uri'
 
-Ingredient.destroy_all
+Ingredient.destroy_all if Rails.env.development?
 
 url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
 parsed = JSON.parse(URI.open(url).read)
 parsed.values.flatten.each do |ingred|
   Ingredient.create(name: ingred.values.first)
 end
+
+
 
 # Ingredient.create(name: "lemon")
 # Ingredient.create(name: "ice")
@@ -21,3 +23,4 @@ end
 # Ingredient.create(name: "port")
 # Ingredient.create(name: "sloe gin")
 # Ingredient.create(name: "cointreau")
+
